requirements.txt
pandas
dataclass-csv
matplotlib
seaborn
statsmodels
sklearn
uvicorn
fastapi
pyyaml